OpenCart_autotests/pyproject.toml
[project]
name = "diplom"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikhail Zhigunov",email = "m.zhigunov@itfbgroup.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "allure-pytest (>=2.13.5,<3.0.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "faker (>=37.0.0,<38.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytest-xdist (>=3.6.1,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "diplom", from = "src"}]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
